Move Navbar component to options page
