+++ src/options/Navbar.module.scss
